Move new_data_processing to data_processing
